build: use TARNAME instead of NAME for paths
PACKAGE_TARNAME is the same as PACKAGE_NAME but normalized, so it should
be safer to use in paths.  For example, on a downstream project, if
spaces or shell metacharacters are added to the package name, a path
that uses PACKAGE_TARNAME should keep working.

From the manual of GNU Autoconf (version 2.69):

>  -- Macro: AC_INIT (PACKAGE, VERSION, [BUG-REPORT], [TARNAME], [URL])
>      Process any command-line arguments and perform initialization and
>      verification.
>
>      Set the name of the PACKAGE and its VERSION.  These are typically
>      used in '--version' support, including that of 'configure'.  The
>      optional argument BUG-REPORT should be the email to which users
>      should send bug reports.  The package TARNAME differs from
>      PACKAGE: the latter designates the full package name (e.g., 'GNU
>      Autoconf'), while the former is meant for distribution tar ball
>      names (e.g., 'autoconf').  It defaults to PACKAGE with 'GNU '
>      stripped, lower-cased, and all characters other than
>      alphanumerics and underscores are changed to '-'.

Note also that by default (on autoconf v2.69), `docdir=@docdir@` in
config.mk.in expands to the following in config.mk:

    docdir=${datarootdir}/doc/${PACKAGE_TARNAME}