ci: sort items on paths-ignore lists
Leave quotes only on paths with wildcards so that they are sorted before
normal paths, then sort everything.

Note that in the current workflow files, ignored directories always use
wildcards, so this also ensures that ignored directories (along with all
other paths with wildcards) are always listed before ignored files
(similarly to `--group-directories-first` in GNU `ls`).

This order is similar to the one on
.github/workflows/profile-checks.yml.