makefiles: deduplicate main target name into new SO var
Put the main target name into a new SO variable, put SO into a new
TARGET variable, make "all" depend on `$(TARGET)` and replace every
other occurrence of the main target name with `$(SO)`.

On the makefiles that build shared objects, to make them more similar.
With this commit, all of their targets are identical.