ci: ignore man/vim paths in build-extra/codeql workflows
Ignore the following paths:

- 'contrib/vim/**'
- 'src/man/*.txt'

To avoid running CI unnecessarily.

Note: build-extra essentially only builds with clang and runs linters on
the C code, so changes to the man pages and vim files should not affect
the result.

Do not ignore them on .github/workflows/build.yml because they are part
of DISTFILES (see the root Makefile).

Misc: I noticed this on #5296 (man pages).