.git-blame-ignore-revs: change format to hash then info
Command used to generate the entries in the new format:

    $ TZ=UTC0 git show --date='format-local:%Y-%m-%d' \
      --pretty='%H # %cd | %s' -s f43382f1e fe0f975f4

Put the committer date ("%cd") and commit message subject ("%s") as a
comment right after the commit hash ("%H").  The pipe ("|") is intended
to improve readability.

This format looks more similar to the output of `git log --oneline` and
should make it easier to visually parse long lists.  It also allows more
hashes to fit into the same amount of lines (at the cost of longer
lines).

Use committer date ("%cd") instead of author date ("%ad") as the former
tells when a commit actually landed on master.  This usually matches the
topological order, which should make it easier to verify that all of the
relevant commits are listed (and that there are no extraneous commits).

Use --date + "%cd" instead of just "%ci" because the former can be used
to ensure that the date is always printed in the same timezone (UTC).