.git-blame-ignore-revs: add Landlock v1 commits
Commands used to find and print the relevant commits:

    $ TZ=UTC0 git log --date='format-local:%Y-%m-%d' \
      --pretty='%H # %cd | %s' 60db9f785~1..60db9f785 |
      grep 'Revert "';
      git log --reverse --pretty=%b 60db9f785~1..60db9f785 |
      sed -E -n 's/^This reverts commit ([0-9a-z]+).*/\1/p' |
      TZ=UTC0 xargs git show --date='format-local:%Y-%m-%d' \
      --pretty='%H # %cd | %s' -s | grep -v 'Merge pull request';
      TZ=UTC0 git log --no-merges --date='format-local:%Y-%m-%d' \
      --pretty='%H # %cd | %s' 54cb3e741~1..54cb3e741

Explanation: The first `git log` basically takes the revision range from
one commit before a given merge commit until the merge commit itself and
prints the commits in that range (which should usually mean all commits
that were in the branch that was merged).  In this case, it's the
commits that do the revert.

The second `git log` finds the hash of all commits that were reverted
and prints them.

The `grep -v` and third `git log` are only needed because the merge
commit of the original branch was reverted directly (on commit
97874c3bf), rather than reverting each individual commit on that branch.
So these commands are used to print all of the commits in the original
branch.

Relates to #5315 #5347.