sort.py: use script name in usage/main docstring
With this, the help section remains consistent regardless of how the
script is called and even if the filename is changed.  For example, if
someone renames "sort.py" to "firejail-sort" and puts it somewhere in
`$PATH`.

Example outputs of the script name (using `print(argv[0]); return`):

    $ ./contrib/sort.py
    ./contrib/sort.py
    $ python contrib/sort.sh
    contrib/sort.py
    $ (cd contrib && ./sort.py)
    ./sort.py

Note: This depends on `os.path` and `sys.argv`, so the imports have to
appear before the docstring.  In which case, the docstring has to be
explicitly assigned to `__doc__` (as it ceases to be the first statement
in the file).

Note2: When running `pydoc ./contrib/sort.py`, `argv[0]` becomes
"/usr/bin/pydoc" (using python 3.10.8-1 on Artix Linux).