makefiles: expand HDRS, SRCS and OBJS immediately
Use immediate expansion of the right-hand side (with `:=`) to set the
variables to the output of the commands rather than to the (text of the)
commands themselves.

This should prevent deferred/lazy evaluation, which is something that
might potentially result in the relevant files being looked up each time
that HDRS and SRCS are evaluated.

Commands used to search and replace:

    git grep -Ilz '^SRCS' -- src | xargs -0 -I '{}' \
        sh -c "printf '%s\n' \"\$(sed \
            -e 's/^HDRS =/HDRS :=/' \
            -e 's/^SRCS =/SRCS :=/' \
            -e 's/^OBJS =/OBJS :=/' '{}')\" >'{}'"