Allow KeePassXC access to its socket path
KeePassXC now exposes its browser socket in a dedicated directory, so
that this directory can be mounted inside the same sandbox as browser
and browsers can talk to it.

This allows using the browser extension when KeePassXC and/or the
browser are sandboxed.

See: https://github.com/keepassxreboot/keepassxc/issues/8018
Fixes: https://github.com/netblue30/firejail/issues/5447