RELNOTES: add new profiles to 0.9.72
Commands used to find the profiles and print the RELNOTES items:

    $ git log --reverse --pretty= --name-only --diff-filter=A \
      0.9.70..0.9.72 -- etc/inc etc/net etc/profile-* | cut -f 3 -d / |
      sed -E -e 's/^([^.]+)\.profile$/\1/' -e 's/$/,/' | tr '\n' ' ' |
      fold -s -w 61 | sed 's/^/  * new profiles: /'; echo

Based on the commands from commit a320957a1 ("RELNOTES: add missing new
profiles", 2022-06-09) / PR #5184.

Note: 61 is used in fold because it's 79 (the default `textwidth` / `tw`
in vim) minus 18 (the length of "  * new profiles: ").

Note2: ".profile" is only trimmed if it's the only suffix, to make it
clear that a new etc/profile-a-l/foo.inc.profile is not a new
etc/inc/foo.inc profile.

Note3: Keep the commas at the end because removing them could need
another `fold` to make the output exactly equivalent to
writing/formatting the items manually.

Note4: There were no profiles removed in 0.9.72:

    $ git log --reverse --pretty= --name-only --diff-filter=D \
      0.9.70..0.9.72 -- etc/inc etc/net etc/profile-*
    $