zoom.profile: whitelist ~/.config/zoom.conf (#5596)
* zoom.profile: whitelist ~/.config/zoom.conf

With Zoom version 5.12.6, Zoom changed how they handle encrypting the local
database. This change resulted in the new file zoom.conf being used. As it is
not allowed by the current profile, this could lead to users losing their chat
history if they cannot be retrieved from the cloud (e.g. when e2e encryption is
used).

* zoom.profile: noblacklist ~/.config/zoom.conf

Additional blacklisting for other programs to the configuration file.