gnome-logs: fix missing machine-id in private-etc
The app can still read the logs without machine-id, but it does show a
large red notification widget stating that the current user doesn't have
the proper privileges to read the logs (even though the user is in the
systemd-journal group).

It was removed on commit 5d0822c52 ("private-etc: big profile changes",
2023-02-05).

Relates to #6400.