Blacklist sendgmail config
sendgmail is a cli tool by Google that "uses Gmail in order to mimic
sendmail for git send-email" as per its own description. In other words it
is a basic sendmail replacement with OAuth2 support to send emails from
Gmail accounts.

https://github.com/google/gmail-oauth2-tools/tree/master/go/sendgmail

Config files location depends on "xdg" build tag. Without the tag it would
be "~/.sendgmail.*". With the tag it is either under
"$XDG_CONFIG_HOME/sendgmail" if set or "~/.config/sendgmail" otherwise.