ci: deb: remove redundant --prefix arguments
The "deb" target depends on the "dist" target, which creates an archive
from DISTFILES.

The arguments to ./configure are misleading, as they do not affect the
archive that is used by `make deb`.  That is the case because the
configure output files (config.mk and config.sh) are not copied into the
dist archive, only their input files (config.mk.in and config.sh.in).

In order to affect the .deb package, the configure arguments have to be
passed to mkdeb.sh, which then forwards them to ./configure itself.

Note: This does not apply to the rpm-based jobs, as `make rpms` uses the
files directly rather than using the dist archive.

Relates to #5154.