Fix space before/after tab in indent
This fixes all of the "space before tab in indent" errors raised by git:

    $ git diff --check 4b825dc642cb6eb9a060e54bf8d69288fbee4904..HEAD |
      grep '^[^+]' | cut -f 3 -d : | LC_ALL=C sort | uniq -c
         72  space before tab in indent.

Commands used to find the errors:

    $ git diff --check 4b825dc642cb6eb9a060e54bf8d69288fbee4904..HEAD
    $ git grep -In "$(printf '\t')     "

Note: Unlike "space before tab in indent", the reverse ("space after tab
in indent") is not reported by git.  That is because spaces could be
intentionally used for alignment or line continuation, but in some cases
they are being used for indentation together with tabs and in others the
formatting is misaligned.  The second command was used to help find and
fix these other issues.