Fix inconsistent leading spaces
Changes:

* Fix spaces being used for indentation in some lines in C
* Remove leading spaces before some goto labels
* Remove leading spaces before the start of some multiline comments
* Change leading spaces to tabs in some multiline macros
* Add missing asterisk to some multiline comments (to match other
  multiline comments and because they are false positives in the
  commands below)

Note: Leading spaces can be used for alignment (such as in function
parameters and function arguments in C) and for line continuation (such
as in long commands in shell scripts).  However, in the above changes
the leading spaces are used for other reasons and do not seem to fit
with the style used.

Commands used to search for errors:

    $ git grep -In '^ [^*]' | grep -E -v \
        -e '(COPYING|README|RELNOTES|configure(.ac)?):' \
        -e '^[^:]+.(md|yml|py):' -e '(bash|zsh)_completion/' \
	-e '^contrib/syntax/' -e '^etc/templates/.*\.txt:' -e '^m4/' \
        -e '^platform/debian/' -e '^src/man/.*\.txt:' \
        -e '.*mkrpm.sh:' -e '.*extract_errnos.sh:'