.gitattributes: mark some binary files as such
Git currently correctly detects them as binary; the changes are done to
avoid depending on the auto-detection and also for documentation.

Commands used to list all of the files that git detects as non-text
files:

    $ git ls-files --eol | grep -e 'i/-text' -e 'w/-text'
    i/-text w/-text attr/text=auto eol=lf   etc-fixes/seccomp-join-bug/eecf35c-backports.zip
    i/-text w/-text attr/text=auto eol=lf   test/appimage/Leafpad-0.8.17-x86_64.AppImage
    i/-text w/-text attr/text=auto eol=lf   test/appimage/Leafpad-0.8.18.1.glibc2.4-x86_64.AppImage
    i/-text w/-text attr/text=auto eol=lf   test/filters/memwrexe
    i/-text w/-text attr/text=auto eol=lf   test/filters/memwrexe-32
    i/-text w/-text attr/text=auto eol=lf   test/filters/namespaces
    i/-text w/-text attr/text=auto eol=lf   test/filters/namespaces-32

Note: The committed seccomp filters do not have a file extension, so
ignore them for now.