pull_request_template.md: restore indentation
This partially reverts commit 375468008 ("docs: remove indents on
top-level lists and tables", 2023-02-01) from PR #5674.

Commands used to undo the changes:

    $ f=.github/pull_request_template.md; \
      git show 3754680087~1:"$f" >"$f"

I had assumed that a blank line after a list item would end the list
(and so I was confused by the amount of indentation used), but that is
apparently not the case.  See the file rendered before/after the
commit[1] [2].

Relates to #2784.

Reported by @rusty-snake[3].

[1] https://github.com/netblue30/firejail/blob/f5d8d8cc7af8f8816c47623515babcefceb7e22f/.github/pull_request_template.md
[2] https://github.com/netblue30/firejail/blob/37546800876d977d77cc86d9b307f8cfa714c1dd/.github/pull_request_template.md
[3] https://github.com/netblue30/firejail/pull/5674#discussion_r1117892922